requirements.txt
discord.py==1.7.2
gspread 
oauth2client